Changing netReadAll to return size_t instead of int, which works better on files bigger than 2 gig.

+++ src/inc/net.h
@@ -47,7 +47,7 @@ FILE *netFileFromSocket(int socket);
 void netBlockBrokenPipes();
 /* Make it so a broken pipe doesn't kill us. */
 
-int netReadAll(int sd, void *vBuf, size_t size);
+size_t netReadAll(int sd, void *vBuf, size_t size);
 /* Read given number of bytes into buffer.
  * Don't give up on first read! */
